backend/requirements.txt
# FastAPI Framework
fastapi>=0.104.1,<0.115.0
uvicorn[standard]>=0.24.0,<0.30.0

# Data validation
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0

# Database
sqlalchemy>=2.0.0,<3.0.0
psycopg2-binary>=2.9.0,<3.0.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.5

# AI Agent
pydantic-ai>=0.0.13

# Testing
pytest>=7.0.0,<9.0.0
